[Chore] Split jobs.rs into seven submodules
`jobs.rs` mixed the public manager surface, the in-memory job map, every
snapshot transition, the planning-to-execution path, progress consumption,
failure classification and the 3D child job in one 1 082-line file. Each
now has its own:

    failure    message -> typed failure                (leaf)
    progress   consuming updates from an execution     (leaf)
    store      the job map, pruning, region locks      (leaf)
    state      every transition and its guards         -> failure progress
    haruki_3d  the 3D child job                        -> progress state store
    runner     planning through execution              -> the rest
    manager    the public surface                      -> progress runner store

`RegionLockMap` and `acquire_region_lock` sit in `store` rather than
`manager`: the runner takes the lock and the manager owns the map, so
keeping them in `manager` made the two import each other. They belong
with the map they guard, and the result is a DAG -- a cycle check over
the `use super::` edges reports none.

Pure move; 33 items in, 33 out, plus the three the split introduced. The
edits are visibility and imports only.

The six tests from the previous commit move to what they cover: cancel
semantics with the manager, the completion-versus-cancel race with the
transitions, pruning and region locks with the store, and keyword
priority with the classifier. Their fixtures move to `test_support`
rather than being copied. Same 188 tests before and after.
